Finished quiz check questions query

+++ eQuiz/modules/eQuiz.Web/Areas/Admin/Models/QuestionBase.cs
@@ -9,13 +9,13 @@ namespace eQuiz.Web.Areas.Admin.Models
     {
         public int Id { get; set; }
         public string Type { get; set; }
-        public int MaxScore { get; set; }
+        public byte MaxScore { get; set; }
         public int UserScore { get; set; }
         public string QuestionText { get; set; }
         public short? Order { get; set; }
 
         // Base constructor
-        public QuestionBase (int id, int maxScore, int userScore, string questionText, short? order)
+        public QuestionBase (int id, byte maxScore, int userScore, string questionText, short? order)
         {
             Id = id;
             MaxScore = maxScore;
